backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
python-multipart==0.0.7
SQLAlchemy==2.0.36
alembic==1.13.3
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0
